+++ server/lib/coflux/orchestration/epoch.ex
@@ -1979,10 +1979,10 @@ defmodule Coflux.Orchestration.Epoch do
   defp ensure_pool_definition(_source_db, _target_db, nil), do: nil
 
   defp ensure_pool_definition(source_db, target_db, old_id) do
-    {:ok, {hash, launcher_id, provides_tag_set_id, accepts_tag_set_id}} =
+    {:ok, {hash, launcher_id, provides_tag_set_id, accepts_tag_set_id, idle_timeout}} =
       query_one!(
         source_db,
-        "SELECT hash, launcher_id, provides_tag_set_id, accepts_tag_set_id FROM pool_definitions WHERE id = ?1",
+        "SELECT hash, launcher_id, provides_tag_set_id, accepts_tag_set_id, idle_timeout FROM pool_definitions WHERE id = ?1",
         {old_id}
       )
 
@@ -2000,7 +2000,8 @@ defmodule Coflux.Orchestration.Epoch do
             hash: {:blob, hash},
             launcher_id: ensure_launcher(source_db, target_db, launcher_id),
             provides_tag_set_id: ensure_tag_set(source_db, target_db, provides_tag_set_id),
-            accepts_tag_set_id: ensure_tag_set(source_db, target_db, accepts_tag_set_id)
+            accepts_tag_set_id: ensure_tag_set(source_db, target_db, accepts_tag_set_id),
+            idle_timeout: idle_timeout
           })
 
         # Copy pool_definition_modules